pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Products API service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "africastalking>=1.2.9",
    "django>=5.2.2",
    "django-pyoidc>=1.0.4",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "uvicorn>=0.34.3",
    "uvicorn-worker>=0.3.0",
]

[dependency-groups]
dev = [
    "aiosmtpd>=1.4.6",
    "ipdb>=0.13.13",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
]

[tool.coverage.run]
source = ["api", "nexus"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "oidc.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pytest.ini_options]
addopts = "--cov=api --cov-report=term-missing --cov-report=xml"
markers = [
    "django_db: mark a test as requiring the Django database",
]
